Refactor for extensibility
Chattie has been extensively refactored. Originally it had many design flaws
that were a direct result of the "weekend powered by energy drinks" that spawned
it. Now it has been refactored to remove any two way referencing pointers (i.e.
the bot and connector both have a pointer to the other) which allows us to
increase speed by invisibly making Chattie concurrent. Additionally it allows
connectors to implement more advanced features. Previously some backends, like
Telegram, offer additional integrations for bots like slash commands or
REPL-esque features. Chattie was unable to take advantage of this because of the
way it was architected. Now the bot class does 0 impure operations (read IO) in
regards to sending or receiving messages. Now the bot simply builds a list of
strings which are the appropriate responses to either the commands called or
handlers triggered. Additionally, the Connector now has full access to the Bot
so it can access it's inventory for persistence or cheat it's way to calling
commands directly.

Another large change is that inventories are now pluggable. Originally Chattie
would store it's inventory in a JSON file and use internal memory to manage the
live state. Now there is an Inventory class and it can be implemented in any way
you like. See `src/chattie/inventory/__init__.py` for a reference class. Chattie
will default to using the json inventory which is the old inventory changed to
this method.