Change the bot, trick, connector, handler contract
This makes it so the Connector is responsible for determining whether
a command or handler should be run. The bot is only responsible for
managing which tricks go with which commands and what the handlers
are. This allows for tighter integration with a connector's backend by
pushing this responsibility closer to that backend.